+++ CHANGELOG.md
@@ -7,7 +7,19 @@ and this project tries to adhere to [Semantic Versioning](https://semver.org/spe
 ## [Unreleased]
 ### Added
 - Timeseries plot (from https://plotly.com/python/time-series/#time-series-with-range-selector-buttons)
+- button to reload scenarios
+- pinned requirements
+- loading spinner
+- messaging to handle warnings and errors
+- corrupt timeseries dates are fixed and user gets warned
 - Radar plot for scalars
 
 ### Changed
 - multiple plots with different options can be loaded
+
+### Fixed
+- added redis to requirements
+- error message if daterange index is wrong
+
+### Cleaned
+- moved timeseries duplicate and columns handling to preprocessing